Simplified Directus base class
